background: when using `gulp.dest` then files will be saved under:     [value returned  from the callback] + file.relativePath.
Wrong assumption:
before this change, the assumption was that it saved as:
    [value returned  from the callback] + file's name

relative path! relative to what?
relative path is based on the glob pattern, so if it was `src/**/*.js` and the file that was extracted out of it was: `src/a.txt` then relative path will be a.txt, and if the file was src/a/b.txt then relative path will be `a/b.txt`. So it is relative to what came before '**'.

Solution:
To handle this, we need to handle 2 different cases - files that are at the root and files that are under src.
for files at the root we need to return:
   [file.cwd] + /out
for files under /src we need to return:
   [file.cwd] + /out/src

to see if a file is under /src we check if the original file has the substring /src at it.